extract global state to MainViewModel class
