+++ androidApp/src/main/java/com/keller/yourpet/androidApp/main/view/MainActivity.kt
@@ -1,10 +1,16 @@
-package com.keller.yourpet.androidApp
+package com.keller.yourpet.androidApp.main.view
 
 import android.os.Bundle
+import androidx.activity.ComponentActivity
 import androidx.activity.compose.setContent
 import androidx.appcompat.app.AppCompatActivity
+import androidx.compose.foundation.isSystemInDarkTheme
 import androidx.compose.material.Text
 import androidx.compose.runtime.Composable
+import androidx.compose.runtime.collectAsState
+import androidx.compose.runtime.getValue
+import androidx.compose.ui.platform.LocalContext
+import androidx.hilt.navigation.compose.hiltViewModel
 import androidx.navigation.compose.NavHost
 import androidx.navigation.compose.composable
 import androidx.navigation.compose.rememberNavController
@@ -12,7 +18,9 @@ import com.keller.yourpet.androidApp.Navigation.Companion.ARG_PET
 import com.keller.yourpet.androidApp.Navigation.Companion.ROUTE_HOME
 import com.keller.yourpet.androidApp.Navigation.Companion.ROUTE_PET_DETAILS
 import com.keller.yourpet.androidApp.home.view.HomeScreen
+import com.keller.yourpet.androidApp.main.viewmodel.MainViewModel
 import com.keller.yourpet.androidApp.petdetails.PetDetailsScreen
+import com.keller.yourpet.androidApp.main.viewmodel.isDark
 import com.keller.yourpet.androidApp.ui.YourPetUITheme
 import dagger.hilt.android.AndroidEntryPoint
 import kotlinx.serialization.decodeFromString
@@ -33,15 +41,20 @@ class MainActivity : AppCompatActivity() {
 
     @Composable
     fun ComposeNavigation() {
+        val viewModel: MainViewModel = hiltViewModel(LocalContext.current as ComponentActivity)
+        val state by viewModel.uiState.collectAsState()
         val navController = rememberNavController()
-        NavHost(navController = navController, startDestination = ROUTE_HOME) {
-            composable(ROUTE_HOME) {
-                HomeScreen(navController) // TODO replace with home vm
-            }
-            composable(ROUTE_PET_DETAILS) {
-                navController.previousBackStackEntry?.arguments?.getString(ARG_PET)?.let {
-                    PetDetailsScreen(pet = Json.decodeFromString(it))
-                } ?: run { Text("Something went wrong") }
+
+        YourPetUITheme(darkTheme = state.themeColor.isDark(isSystemInDarkTheme())) {
+            NavHost(navController = navController, startDestination = ROUTE_HOME) {
+                composable(ROUTE_HOME) {
+                    HomeScreen(navController)
+                }
+                composable(ROUTE_PET_DETAILS) {
+                    navController.previousBackStackEntry?.arguments?.getString(ARG_PET)?.let {
+                        PetDetailsScreen(pet = Json.decodeFromString(it))
+                    } ?: run { Text("Something went wrong") }
+                }
             }
         }
     }
