+++ androidApp/src/main/java/com/keller/yourpet/androidApp/settings/view/SettingsScreen.kt
@@ -1,5 +1,6 @@
 package com.keller.yourpet.androidApp.settings.view
 
+import androidx.activity.ComponentActivity
 import androidx.compose.foundation.layout.Box
 import androidx.compose.foundation.layout.Column
 import androidx.compose.foundation.layout.Row
@@ -20,17 +21,18 @@ import androidx.compose.runtime.remember
 import androidx.compose.runtime.setValue
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
+import androidx.compose.ui.platform.LocalContext
 import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.window.PopupProperties
 import androidx.hilt.navigation.compose.hiltViewModel
-import com.keller.yourpet.androidApp.settings.viewmodel.SettingsViewModel
-import com.keller.yourpet.androidApp.settings.viewmodel.ThemeColor
+import com.keller.yourpet.androidApp.main.viewmodel.MainViewModel
+import com.keller.yourpet.androidApp.main.viewmodel.ThemeColor
 import com.keller.yourpet.androidApp.ui.UIGroups
 
 @Composable
 fun SettingsScreen() {
-    val viewModel: SettingsViewModel = hiltViewModel()
-    val state by viewModel.uiState.collectAsState()
+    val mainViewModel: MainViewModel = hiltViewModel(LocalContext.current as ComponentActivity)
+    val uiState by mainViewModel.uiState.collectAsState()
     var themeMenuExpanded by remember { mutableStateOf(false) }
 
     Column {
@@ -42,7 +44,7 @@ fun SettingsScreen() {
                     .wrapContentSize(Alignment.TopEnd)
             ) {
                 Row(modifier = Modifier) {
-                    Text(text = state.themeColor.toString())
+                    Text(text = uiState.themeColor.toString())
                     IconButton(onClick = { themeMenuExpanded = true }) {
                         Icon(Icons.Default.MoreVert, contentDescription = "Theme")
                     }
@@ -55,7 +57,7 @@ fun SettingsScreen() {
                     ThemeColor.values().forEach {
                         DropdownMenuItem(
                             onClick = {
-                                viewModel.onThemeColorSelected(it)
+                                mainViewModel.onThemeColorSelected(it)
                                 themeMenuExpanded = false
                             },
                         ) {
