--- androidApp/src/main/java/com/keller/yourpet/androidApp/settings/viewmodel/SettingsViewState.kt
@@ -1,9 +1,0 @@
-package com.keller.yourpet.androidApp.settings.viewmodel
-
-data class SettingsViewState(val themeColor: ThemeColor)
-
-enum class ThemeColor {
-    LIGHT,
-    DARK,
-    SYSTEM
-}
